backend/requirements.txt
certifi==2025.1.31
fastapi==0.115.11
langchain_core==0.2.43
langchain-openai==0.1.25
langgraph_core==0.2.43
openai==1.65.4
pydantic==2.10.6
requests==2.31.0
tavily_python==0.5.1
uvicorn[standard]==0.34.0
websockets==12.0
python-dotenv==1.0.1 